+++ backend/src/user/dto/send-verification-email.dto.ts
@@ -1,5 +1,4 @@
 import { ApiProperty } from "@nestjs/swagger";
-export class SendVerificationEmailDto {
 
 export class SendVerificationEmailDto {
   @ApiProperty({