+++ CHANGELOG.md
@@ -1,4 +1,25 @@
 
+# v3.0.0 (18.04.2022)
+
+## Features:
+
+Frontend:
+1. Finished profile. Removed cors from backend
+
+## Features:
+
+Frontend:
+1. Added updated profile component
+2. Validator,
+3. Updated profile saga
+
+
+# v2.9.9 (18.04.2022)
+
+## Features:
+
+Frontend:
+1. Deleted unnecessary files, updated messages types
 
 # v2.9.8 (18.04.2022)
 
@@ -7,7 +28,7 @@
 Frontend:
 1. Updated forgot password functional
 2. Added modal window when password was updated
-3. 
+
 # v2.9.7. (15.04.2022)
 
 ## Features: